Moved tests so they're in root folder rather than in script's folder (makes them eaiser to run on travis and in general).
