+++ package.json
@@ -77,6 +77,7 @@
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^5.1.1",
     "css-loader": "^3.3.2",
+    "del": "^5.1.0",
     "eslint": "^6.7.2",
     "eslint-config-wordpress": "^2.0.0",
     "eslint-plugin-import": "^2.19.1",
@@ -85,6 +86,7 @@
     "eslint-plugin-react": "^7.17.0",
     "eslint-plugin-wordpress": "^0.1.0",
     "file-loader": "^5.0.2",
+    "fs-extra": "^8.1.0",
     "husky": "^3.1.0",
     "import-glob-loader": "^1.1.0",
     "jquery": "^3.4.1",
@@ -96,6 +98,7 @@
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "postcss-loader": "^3.0.0",
     "precss": "^4.0.0",
+    "promisify-child-process": "^3.1.3",
     "raw-loader": "^4.0.0",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
@@ -123,11 +126,9 @@
     "@storybook/addons": "^5.2.8",
     "@storybook/react": "^5.2.8",
     "@storybook/storybook-deployer": "^2.8.1",
-    "del": "^5.1.0",
     "docpress": "0.8.0",
-    "fs-extra": "^8.1.0",
     "jest": "^25.1.0",
-    "promisify-child-process": "^3.1.3",
+    "replace-in-file": "^5.0.2",
     "sassdoc": "^2.7.1"
   },
   "docpress": {
