+++ tests/unit/create-wp-project/basics/misc.test.js
@@ -3,7 +3,7 @@
  *
  * @group unit
  */
-const { capCase } = require('../../../src/basics/misc.js');
+const { capCase } = require('../../../../setup/create-wp-project/src/basics/misc.js');
 
 test('Should make cap case of: aaa aaa', () => {
   expect(capCase('aaa aaa')).toBe('Aaa_Aaa');
